LibRuntime: Move a few more stdio functions from LibC
