+++ Userland/Libraries/LibRuntime/Serenity/CMakeLists.txt
@@ -15,7 +15,7 @@ set(LIBRUNTIME_SOURCES
 )
 
 # Prevent GCC from removing null checks by marking the `FILE*` argument non-null
-set_source_files_properties(LibC/stdio.cpp PROPERTIES COMPILE_FLAGS -fno-builtin-fwrite)
+set_source_files_properties(LibC/stdio.cpp PROPERTIES COMPILE_FLAGS "-fno-builtin-fwrite -fno-builtin-fputc -fno-builtin-fputs")
 # Prevent naively implemented string functions (like strlen) from being "optimized" into a call to themselves.
 set_source_files_properties(LibC/string.cpp PROPERTIES COMPILE_FLAGS -fno-builtin)
 set_source_files_properties(PossiblyThrowingCallback.cpp PROPERTIES COMPILE_FLAGS -fexceptions)
